+++ packages/interactive_media_ads/android/src/main/kotlin/dev/flutter/packages/interactive_media_ads/AdsManagerProxyApi.kt
@@ -14,6 +14,10 @@ import com.google.ads.interactivemedia.v3.api.AdsManager
  */
 class AdsManagerProxyApi(override val pigeonRegistrar: ProxyApiRegistrar) :
     PigeonApiAdsManager(pigeonRegistrar) {
+  override fun adCuePoints(pigeon_instance: AdsManager): List<Double> {
+    return pigeon_instance.adCuePoints.map { it.toDouble() }
+  }
+
   override fun discardAdBreak(pigeon_instance: AdsManager) {
     pigeon_instance.discardAdBreak()
   }
@@ -24,5 +28,13 @@ class AdsManagerProxyApi(override val pigeonRegistrar: ProxyApiRegistrar) :
 
   override fun start(pigeon_instance: AdsManager) {
     pigeon_instance.start()
+  }
+
+  override fun resume(pigeon_instance: AdsManager) {
+    pigeon_instance.resume()
+  }
+
+  override fun skip(pigeon_instance: AdsManager) {
+    pigeon_instance.skip()
   }
 }