+++ packages/interactive_media_ads/test/android/ads_manager_test.mocks.dart
@@ -393,6 +393,13 @@ class MockAdEventListener extends _i1.Mock implements _i2.AdEventListener {
 /// See the documentation for Mockito's code generation for more information.
 class MockAdsManager extends _i1.Mock implements _i2.AdsManager {
   @override
+  List<double?> get adCuePoints => (super.noSuchMethod(
+        Invocation.getter(#adCuePoints),
+        returnValue: <double?>[],
+        returnValueForMissingStub: <double?>[],
+      ) as List<double?>);
+
+  @override
   _i2.PigeonInstanceManager get pigeon_instanceManager => (super.noSuchMethod(
         Invocation.getter(#pigeon_instanceManager),
         returnValue: _FakePigeonInstanceManager_0(
@@ -429,6 +436,26 @@ class MockAdsManager extends _i1.Mock implements _i2.AdsManager {
   _i4.Future<void> start() => (super.noSuchMethod(
         Invocation.method(
           #start,
+          [],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+
+  @override
+  _i4.Future<void> resume() => (super.noSuchMethod(
+        Invocation.method(
+          #resume,
+          [],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+
+  @override
+  _i4.Future<void> skip() => (super.noSuchMethod(
+        Invocation.method(
+          #skip,
           [],
         ),
         returnValue: _i4.Future<void>.value(),